Connect platform-shared to web, refactor package structure
