+++ web/src/jsMain/kotlin/io/vitalir/web/Main.kt
@@ -1,4 +1,6 @@
-import io.vitalir.web.main.MainPage
+package io.vitalir.web
+
+import io.vitalir.web.pages.main.MainPage
 import org.jetbrains.compose.web.renderComposable
 
 fun main() {
