+++ web/src/jsMain/kotlin/io/vitalir/web/pages/main/models/Repository.kt
@@ -1,4 +1,4 @@
-package io.vitalir.web.main.models
+package io.vitalir.web.pages.main.models
 
 data class Repository(
     val name: String,
