+++ web/src/jsMain/kotlin/io/vitalir/web/pages/main/views/MainContentView.kt
@@ -1,28 +1,14 @@
-package io.vitalir.web.main.views
+package io.vitalir.web.pages.main.views
 
 import androidx.compose.runtime.Composable
-import io.vitalir.web.main.models.Repository
+import io.vitalir.web.pages.main.models.Repository
 import org.jetbrains.compose.web.css.DisplayStyle
 import org.jetbrains.compose.web.css.FlexDirection
-import org.jetbrains.compose.web.css.borderWidth
-import org.jetbrains.compose.web.css.color
 import org.jetbrains.compose.web.css.cssRem
 import org.jetbrains.compose.web.css.display
 import org.jetbrains.compose.web.css.flexDirection
-import org.jetbrains.compose.web.css.fontSize
-import org.jetbrains.compose.web.css.fontWeight
-import org.jetbrains.compose.web.css.gap
-import org.jetbrains.compose.web.css.lineHeight
-import org.jetbrains.compose.web.css.marginTop
 import org.jetbrains.compose.web.css.padding
-import org.jetbrains.compose.web.css.px
-import org.jetbrains.compose.web.css.rgb
-import org.jetbrains.compose.web.css.textAlign
-import org.jetbrains.compose.web.dom.A
 import org.jetbrains.compose.web.dom.Div
-import org.jetbrains.compose.web.dom.P
-import org.jetbrains.compose.web.dom.Span
-import org.jetbrains.compose.web.dom.Text
 
 @Composable
 internal fun MainContentView(
